+++ packages/gestalt/src/MeasurementStore.js
@@ -1,7 +1,8 @@
 // @flow
 import type { Cache } from './Cache.js';
 
-export default class MeasurementStore<T, V> implements Cache<T, V> {
+export default class MeasurementStore<T: {} | $ReadOnlyArray<mixed>, V>
+  implements Cache<T, V> {
   map: WeakMap<T, V> = new WeakMap();
 
   get(key: T): ?V {