+++ packages/gestalt/src/uniformRowLayout.test.js
@@ -25,6 +25,7 @@ test('empty', () => {
     cache: stubCache(),
     width: 500,
   });
+  // $FlowFixMe: new errors found from flow 0.96 upgrade
   expect(layout([])).toEqual([]);
 });
 
@@ -38,6 +39,7 @@ test('one row, equal heights', () => {
     width: 500,
   });
 
+  // $FlowFixMe: new errors found from flow 0.96 upgrade
   expect(layout(['a', 'b', 'c'])).toEqual([
     { top: 0, left: 0, width: 236, height: 100 },
     { top: 0, left: 250, width: 236, height: 100 },
@@ -55,6 +57,7 @@ test('one column, equal heights', () => {
     width: 250,
     minCols: 1,
   });
+  // $FlowFixMe: new errors found from flow 0.96 upgrade
   expect(layout(['a', 'b', 'c'])).toEqual([
     { top: 0, left: 0, width: 236, height: 100 },
     { top: 114, left: 0, width: 236, height: 100 },
@@ -71,6 +74,7 @@ test('one row, unequal heights', () => {
     }),
     width: 500,
   });
+  // $FlowFixMe: new errors found from flow 0.96 upgrade
   expect(layout(['a', 'b', 'c'])).toEqual([
     { top: 0, left: 0, width: 236, height: 100 },
     { top: 0, left: 250, width: 236, height: 120 },
@@ -88,6 +92,7 @@ test('multiple rows, unequal heights', () => {
     }),
     width: 750,
   });
+  // $FlowFixMe: new errors found from flow 0.96 upgrade
   expect(layout(['a', 'b', 'c', 'd'])).toEqual([
     { top: 0, left: 0, width: 236, height: 100 },
     { top: 0, left: 250, width: 236, height: 120 },