+++ CHANGELOG.md
@@ -7,13 +7,54 @@
   </summary>
 
 ### Minor
+
+- Icon: Add new icon of `heart-outline` and update icon of `heart` (#526)
+
 ### Patch
 
 - Typography: Fixed a tiny bug where our CSS file had a few uses of curly quotes instead of regular quotes (#524)
 
 </details>
 
+## 0.97.0 (May 15, 2019)
+
+### Minor
+
+- IconButton: Add new `disabled` prop and stylings to `IconButton` component (#521)
+- Icon: Add new icon of `ads-stats` and `ads-overview` (#522)
+
+## 0.96.0 (May 6, 2019)
+
+## Minor
+
+- Icon: add calendar icon to gestalt (#512)
+- Icon: add lightning bolt icon (#513)
+- Flow: upgrade version to [0.97.0](https://github.com/facebook/flow/releases/tag/v0.97.0) (#515)
+
+## 0.95.0 (April 10, 2019)
+
+### Minor
+
+- Tooltip: remove focus from revealing Tooltip (#506)
+- TextField: Add autocomplete prop value to TextField (#508)
+- Upgrade flow version to [0.96.0](https://github.com/facebook/flow/releases/tag/v0.96.0) (#509)
+
+### Patch
+
+## 0.94.0 (March 25, 2019)
+
+### Minor
+
+- Icon: add some new text related icons (#496)
+- Modal: add a new sizing option to Modal to match Flyout (#499)
+- Modal: add the ability to set a custom header beyond text (#500)
+- Tooltip: introduce new Tooltip wrapper component (#501)
+- Touchable: cleanup event handlers (#502)
+- TextArea/TextField/SelectList: cleanup components and remove gDSFP (#503)
+- Upgrade React version to [16.8.5](https://github.com/facebook/react/releases/tag/v16.8.5) (#504)
+
 ## 0.93.0 (March 13, 2019)
+
 - Mask: add new prop `willChangeTransform` default true which can turn off `willChange:transform` property in CSS (#494)
 
 ## 0.92.0 (March 7, 2019)